pyproject.toml
# Check poetry location by:
# Unix - which poetry
# Windows - where poetry

# poetry install --with dev,test,docs
# poetry run pytest

[tool.poetry]
name = "langgraph-compare"
version = "0.1.2"
description = "Package for comparing and monitoring performance of LangGraph architectures."
authors = ["Tomasz Serafiński <tomasz@serafinski.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langgraph = "^0.2.45"
pandas = "^2.2.3"
pm4py = "^2.7.12.1"
msgpack = "^1.1.0"
jinja2 = "^3.1.2"
langgraph-checkpoint-sqlite = "^2.0.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
langsmith = "^0.2.4"
langchain-openai = "^0.2.6"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
tavily-python = "^0.5.0"
langchain-experimental = "^0.3.3"
matplotlib = "^3.9.2"
langchain-core = "^0.3.15"
pydantic = "^2.9.2"
prefixspan = "^0.5.2"
pyvis = "^0.3.2"
numpy = "<2.0.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.poetry.group.docs]
optional = true

# sphinx wciaz wymaga zainstalowania przez pip?
[tool.poetry.group.docs.dependencies]
# jezeli chcesz zainstalowac zaleznosci do dokumentacji, musisz uzyc min python 3.10
sphinx = { version = "^8.1.3", markers = "python_version >= '3.10'" }
sphinx-rtd-theme = "^3.0.2"
sphinx-copybutton = "^0.5.2"
sphinx-last-updated-by-git = "^0.3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
